Introduce data/tests sub-folder
